fix(product): add nullcheck for non existing products

+++ src/Services/Product/eTenpo.Product.Application/ProductFeature/Delete/DeleteProductHandler.cs
@@ -1,5 +1,6 @@
 using eTenpo.Product.Application.CommandQueryAbstractions;
 using eTenpo.Product.Domain.Contracts;
+using eTenpo.Product.Domain.Exceptions;
 
 namespace eTenpo.Product.Application.ProductFeature.Delete;
 
@@ -16,7 +17,12 @@ public class DeleteProductHandler : ICommandHandler<DeleteProductCommand>
     
     public async Task Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
-        var product = await this.productRepository.GetById(request.Id, cancellationToken);
+        var product = await this.productRepository.FindById(request.Id, cancellationToken);
+
+        if (product is null)
+        {
+            throw new EntityNotFoundException($"Product with id {request.Id} could not be found");
+        }
         
         // mark as deleted in changeTracker
         this.productRepository.Delete(product);
